Add jrisc python module
Wrap the JRISC library functionality in a single
python function: disassemble(). This function does
roughly the same thing as jdis, but as a python
function call operating on raw memory.

A build harness, setup.py, and a Makefile to
wrap calling it and installing the resulting
python module in some path, are also added. There
were two options when writing these:

(1) The Unsupported/deprecated distutils method:

In short, make a call like this:

  $ PYTHONPATH="<jaguar-sdk_dir>/tools/lib/python<major.minor>/site-packages" \
        ./setup.py install --prefix "<jaguar-sdk_dir>/tools"

And then create a file with the following contents:

  import os
  import site
  site.addsitedir(os.path.dirname(os.path.realpath(__file__)))

Called "sitecustomize.py" in <jaguar-sdk_dir/tools/lib/python<major.minor>/site-packages/sitecustomize.py and add that directory to PYTHONPATH.

What's up with that sitecustomize.py file? Well,
python searches all directories in its PYTHONPATH,
but setup.py doesn't install in that path
directly. Instead, it installs a .egg directory in
that path and adds an entry to the
easy-install.pth file in that top-level directory,
which in turn causes python to search the .egg
directory for modules IF PYTHON PARSES THE .PTH
FILE! Python does not, by default, parse these
.pth files outside of its core paths unless
explicitly told to by something like the above
sitecustomize.py file. What a pain.

There also appears to be no straight-forward way
to deduce the versioned directory used above other
than parsing out version numbers.

(2) Use pip: The "supported" install method.

This is what is done here. Something like:

  $ pip install --target=<jaguar-sdk_dir>/tools/lib/python

And then add <jaguar-sdk_dir>/tools/lib/python to
PYTHONPATH. The downside here is that Debian-based
systems seem to carry some patches that break this
in various ways when doing non-trivial installs.
Luckily, it was possible to work around that in
the included helper Makefile

Also note with either method, the installed
library's filename will be tied to the version of
python in use, so it will have to be rebuilt if
the python interpreter installed on the system is
upgraded.